Add support for user defined genetic code
Needed to include an easy way to change the genetic code,
since the program was being used for invertebrate mitochondria.
The new flag, -t, receives a number corresponding to the NCBI genetic code tables.
It will complain if it receives a table that is not mitochondrial. Default table is still
the Vertebrate Mitochondrial Code.