Use the translation from the genebank file before trying to translate
Some genebank files have the translation anotated in the gene feature, and this translation is usually
better than translating just using a table. For example, some mitochondrial genes (such as ND3) may have frame shifts
in their sequences. This is not pratical to implement in this program, so the next best option is to use the translation
avaiable.